feat: separate UI and onSubmit logic for `TaskCreationDrawer`
